move participant dashboard to it's own folder
